renamed common_examples
