+++ integration_tests/examples/audio.rs
@@ -8,7 +8,7 @@ use std::{
     time::Duration,
 };
 
-use integration_tests::examples_common::{
+use integration_tests::examples::{
     self, download_file, start_ffplay, start_websocket_thread, stream_audio, stream_video,
 };
 
@@ -56,7 +56,7 @@ fn start_example_client_code() -> Result<()> {
     let elephant_path = download_file(ELEPHANT_DREAM_FILE_URL, ELEPHANT_DREAM_FILE_PATH)?;
 
     info!("[example] Send register input request.");
-    examples_common::post(
+    examples::post(
         "input/input_1/register",
         &json!({
             "type": "rtp_stream",
@@ -68,7 +68,7 @@ fn start_example_client_code() -> Result<()> {
     )?;
 
     info!("[example] Send register input request.");
-    examples_common::post(
+    examples::post(
         "input/input_2/register",
         &json!({
             "type": "rtp_stream",
@@ -80,7 +80,7 @@ fn start_example_client_code() -> Result<()> {
     )?;
 
     info!("[example] Send register input request.");
-    examples_common::post(
+    examples::post(
         "input/input_3/register",
         &json!({
             "type": "rtp_stream",
@@ -92,7 +92,7 @@ fn start_example_client_code() -> Result<()> {
     )?;
 
     info!("[example] Send register input request.");
-    examples_common::post(
+    examples::post(
         "input/input_4/register",
         &json!({
             "type": "rtp_stream",
@@ -104,7 +104,7 @@ fn start_example_client_code() -> Result<()> {
     )?;
 
     info!("[example] Send register output request.");
-    examples_common::post(
+    examples::post(
         "output/output_1/register",
         &json!({
             "type": "rtp_stream",
@@ -140,7 +140,7 @@ fn start_example_client_code() -> Result<()> {
     )?;
 
     info!("[example] Send register output request.");
-    examples_common::post(
+    examples::post(
         "output/output_2/register",
         &json!({
             "type": "rtp_stream",
@@ -164,7 +164,7 @@ fn start_example_client_code() -> Result<()> {
     std::thread::sleep(Duration::from_millis(500));
 
     info!("[example] Start pipeline");
-    examples_common::post("start", &json!({}))?;
+    examples::post("start", &json!({}))?;
 
     stream_video(IP, INPUT_1_PORT, bunny_path.clone())?;
     stream_audio(IP, INPUT_2_PORT, bunny_path, "libopus")?;
