+++ integration_tests/examples/test_pattern.rs
@@ -5,7 +5,7 @@ use serde::Deserialize;
 use serde_json::json;
 use std::{env, thread};
 
-use integration_tests::examples_common::{
+use integration_tests::examples::{
     self, start_ffplay, start_websocket_thread, stream_ffmpeg_testsrc,
 };
 
@@ -41,7 +41,7 @@ fn start_example_client_code() -> Result<()> {
     start_websocket_thread();
 
     info!("[example] Send register input request.");
-    let RegisterResponse { port: input_port } = examples_common::post(
+    let RegisterResponse { port: input_port } = examples::post(
         "input/input_1/register",
         &json!({
             "type": "rtp_stream",
@@ -57,7 +57,7 @@ fn start_example_client_code() -> Result<()> {
 
     let shader_source = include_str!("./silly.wgsl");
     info!("[example] Register shader transform");
-    examples_common::post(
+    examples::post(
         "shader/example_shader/register",
         &json!({
             "source": shader_source,
@@ -78,7 +78,7 @@ fn start_example_client_code() -> Result<()> {
     });
 
     info!("[example] Send register output request.");
-    examples_common::post(
+    examples::post(
         "output/output_1/register",
         &json!({
             "type": "rtp_stream",
@@ -104,7 +104,7 @@ fn start_example_client_code() -> Result<()> {
     stream_ffmpeg_testsrc(IP, input_port, VIDEO_RESOLUTION)?;
 
     info!("[example] Start pipeline");
-    examples_common::post("start", &json!({}))?;
+    examples::post("start", &json!({}))?;
 
     Ok(())
 }
