+++ integration_tests/examples/text.rs
@@ -4,7 +4,7 @@ use log::{error, info};
 use serde_json::json;
 use std::thread;
 
-use integration_tests::examples_common::{self, start_ffplay, start_websocket_thread};
+use integration_tests::examples::{self, start_ffplay, start_websocket_thread};
 
 const VIDEO_RESOLUTION: Resolution = Resolution {
     width: 1920,
@@ -32,7 +32,7 @@ fn start_example_client_code() -> Result<()> {
     start_websocket_thread();
 
     info!("[example] Send register output request.");
-    examples_common::post(
+    examples::post(
         "output/output_1/register",
         &json!({
             "type": "rtp_stream",
@@ -66,7 +66,7 @@ fn start_example_client_code() -> Result<()> {
     )?;
 
     info!("[example] Start pipeline");
-    examples_common::post("start", &json!({}))?;
+    examples::post("start", &json!({}))?;
 
     Ok(())
 }
