+++ integration_tests/examples/tiles.rs
@@ -4,7 +4,7 @@ use log::{error, info};
 use serde_json::json;
 use std::thread;
 
-use integration_tests::examples_common::{
+use integration_tests::examples::{
     self, start_ffplay, start_websocket_thread, stream_ffmpeg_testsrc,
 };
 
@@ -35,7 +35,7 @@ fn start_example_client_code() -> Result<()> {
     start_websocket_thread();
 
     info!("[example] Send register input request.");
-    examples_common::post(
+    examples::post(
         "input/input_1/register",
         &json!({
             "type": "rtp_stream",
@@ -72,7 +72,7 @@ fn start_example_client_code() -> Result<()> {
     };
 
     info!("[example] Send register output request.");
-    examples_common::post(
+    examples::post(
         "output/output_1/register",
         &json!({
             "type": "rtp_stream",
@@ -96,7 +96,7 @@ fn start_example_client_code() -> Result<()> {
 
     for i in 1..=16 {
         info!("[example] Update output");
-        examples_common::post(
+        examples::post(
             "output/output_1/update",
             &json!({
                 "video": {
@@ -108,14 +108,14 @@ fn start_example_client_code() -> Result<()> {
     }
 
     info!("[example] Start pipeline");
-    examples_common::post("start", &json!({}))?;
+    examples::post("start", &json!({}))?;
 
     info!("[example] Start input stream");
     stream_ffmpeg_testsrc(IP, INPUT_PORT, VIDEO_RESOLUTION)?;
 
     for i in 0..16 {
         info!("[example] Update output");
-        examples_common::post(
+        examples::post(
             "output/output_1/update",
             &json!({
                 "video": {
@@ -127,7 +127,7 @@ fn start_example_client_code() -> Result<()> {
     }
 
     info!("[example] Update output");
-    examples_common::post(
+    examples::post(
         "output/output_1/update",
         &json!({
             "video": {
